refactor extracting/saving commits and do not perform unnecessary autoboxing

+++ coderadar-rest/src/test/java/io/reflectoring/coderadar/rest/unit/analyzerconfig/GetAnalyzerConfigurationControllerTest.java
@@ -32,7 +32,7 @@ class GetAnalyzerConfigurationControllerTest {
         testSubject.getAnalyzerConfiguration(1L);
 
     Assertions.assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
-    Assertions.assertEquals(1L, responseEntity.getBody().getId().longValue());
+    Assertions.assertEquals(1L, responseEntity.getBody().getId());
     Assertions.assertEquals("analyzer", responseEntity.getBody().getAnalyzerName());
     Assertions.assertEquals(true, responseEntity.getBody().getEnabled());
   }
