+++ coderadar-rest/src/test/java/io/reflectoring/coderadar/rest/unit/module/CreateModuleControllerTest.java
@@ -29,6 +29,6 @@ class CreateModuleControllerTest {
     ResponseEntity<IdResponse> responseEntity = testSubject.createModule(command, 5L);
 
     Assertions.assertEquals(HttpStatus.CREATED, responseEntity.getStatusCode());
-    Assertions.assertEquals(1L, responseEntity.getBody().getId().longValue());
+    Assertions.assertEquals(1L, responseEntity.getBody().getId());
   }
 }
